requirements.txt
appdirs==1.4.4
certifi==2021.5.30
chardet==4.0.0
distlib==0.3.2
filelock==3.0.12
idna==2.10
ldap3==2.9
pipenv==2021.5.29
pyasn1==0.4.8
pydirectory==0.1.15
pyTelegramBotAPI==3.8.1
requests==2.25.1
six==1.16.0
speedtest-cli==2.1.3
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4